+++ test/helpers/e2e/fields/array/index.ts
@@ -1,0 +1,4 @@
+export { addArrayRow, addArrayRowBelow } from './addArrayRow.js'
+export { duplicateArrayRow } from './duplicateArrayRow.js'
+export { openArrayRowActions } from './openArrayRowActions.js'
+export { removeArrayRow } from './removeArrayRow.js'
