+++ test/helpers/e2e/fields/blocks/index.ts
@@ -1,0 +1,4 @@
+export { addBlock } from './addBlock.js'
+export { openBlocksDrawer } from './openBlocksDrawer.js'
+export { removeAllBlocks } from './removeAllBlocks.js'
+export { reorderBlocks } from './reorderBlocks.js'
