+++ test/helpers/e2e/fields/blocks/removeAllBlocks.ts
